Cleanup CrystalEdit: `size_t` related issues
 * eliminate conversion warnings for `size_t` related expressions.
 * convert various `int` and `long` to `size_t`
 * add `static_cast<int>` or add `static_cast<long>` as appropriate
 * classical `(int)` casting is used in ASSERT, for simple convenience.
 * add a few ASSERT() statements to assure that various `size_t` values
	remain within the positive `int` value range.
 * Since `-1` is often used as a "size flag", these changes now
	explicitly test for `x != -1` rather than `x >= 0`

While this commit changes various parameters and data items to `size_t`,
the focus of these changes was to completely eliminate the various
warnings. This code is still fundamentally restricted to 2GB of text.

There is still much work to do to generalize the fundamentals of
CrystalEdit to be fully `size_t` capable (for the X64 platform).  Two
items of important note:
 	1. MFC's CPOINT structure is limited to `LONG` values.
	   The `.x` member should become `size_t`.
	2. MFC's CSTRING data is limited to `int` maximum length

+++ Externals/crystaledit/editlib/UndoRecord.h
@@ -28,7 +28,7 @@ private:
   //
   struct TextBuffer
   {
-    int size;
+    size_t size;
     TCHAR data[1];
   };
 
@@ -71,7 +71,7 @@ private:
     delete m_paSavedRevisionNumbers;
   }
 
-  void SetText (LPCTSTR pszText, int cchText);
+  void SetText (LPCTSTR pszText, size_t cchText);
   void FreeText ();
 
   LPCTSTR GetText () const
@@ -84,7 +84,7 @@ private:
     return m_szText;
   }
 
-  int GetTextLength () const
+  size_t GetTextLength () const
   {
     if (((INT_PTR)m_pszText >> 16) != 0)
       return m_pszText->size;
